+++ src/test/java/com/epam/tc/hw3/tests/Ex1FromHW2.java
@@ -1,9 +1,9 @@
-package com.epam.tc.hw3.scripts;
+package com.epam.tc.hw3.tests;
 
 
 import com.epam.tc.hw3.pages.HomePage;
 import com.epam.tc.hw3.pages.LoginPage;
-import com.epam.tc.hw4.scripts.BaseTest;
+import com.epam.tc.hw4.tests.BaseTest;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
