+++ src/test/java/com/epam/tc/hw3/tests/Ex2FromHW2.java
@@ -1,9 +1,9 @@
-package com.epam.tc.hw3.scripts;
+package com.epam.tc.hw3.tests;
 
 import com.epam.tc.hw3.pages.DifferentElementsPage;
 import com.epam.tc.hw3.pages.HomePage;
 import com.epam.tc.hw3.pages.LoginPage;
-import com.epam.tc.hw4.scripts.BaseTest;
+import com.epam.tc.hw4.tests.BaseTest;
 import org.assertj.core.api.SoftAssertions;
 import org.testng.annotations.Test;
