+++ src/test/java/com/epam/tc/hw4/tests/FailedTest.java
@@ -1,4 +1,4 @@
-package com.epam.tc.hw4.scripts;
+package com.epam.tc.hw4.tests;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.Feature;
@@ -7,12 +7,11 @@ import org.testng.annotations.Test;
 
 @Feature("Home page feature")
 @Story("Home page story")
-public class TestHomePage extends BaseTest {
-
+public class FailedTest extends BaseTest {
     @Test
-    @Description("Test home page")
+    @Description("Failed test")
     public void testHomePage() {
-        assertStep.assertTitle("Home Page");
+        assertStep.assertTitle("Home Page1");
         actionStep.login(USER_LOGIN, USER_PASSWORD);
         assertStep.assertLogin(USER_NAME);
         assertStep.assertHeaders();
