+++ src/test/java/com/epam/tc/hw4/tests/TestDifferentElementsPage.java
@@ -1,4 +1,4 @@
-package com.epam.tc.hw4.scripts;
+package com.epam.tc.hw4.tests;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.Feature;
