+++ src/test/java/com/epam/tc/hw4/utils/ActionStep.java
@@ -1,6 +1,6 @@
 package com.epam.tc.hw4.utils;
 
-import static com.epam.tc.hw4.scripts.BaseTest.webDriver;
+import static com.epam.tc.hw4.tests.BaseTest.webDriver;
 
 import io.qameta.allure.Step;
 import org.openqa.selenium.WebDriver;
