+++ src/test/java/com/epam/tc/hw4/utils/AssertStep.java
@@ -1,6 +1,6 @@
 package com.epam.tc.hw4.utils;
 
-import static com.epam.tc.hw4.scripts.BaseTest.webDriver;
+import static com.epam.tc.hw4.tests.BaseTest.webDriver;
 
 import io.qameta.allure.Step;
 import java.util.ArrayList;
