+++ src/test/resources/runners/suiteHW3.xml
@@ -1,8 +1,8 @@
 <suite name="suiteHW3" verbose="1" >
     <test name="Test" >
         <classes>
-            <class name="com.epam.tc.hw3.scripts.Ex1FromHW2"/>
-            <class name="com.epam.tc.hw3.scripts.Ex2FromHW2"/>
+            <class name="com.epam.tc.hw3.tests.Ex1FromHW2"/>
+            <class name="com.epam.tc.hw3.tests.Ex2FromHW2"/>
         </classes>
     </test>
 </suite>
